+++ NEWS.md
@@ -1,1 +1,9 @@
-# sbShinyModules (development version)
+# sbShinyModules (0.0.9000)
+
+Initial package version. 
+
+#sbShinyModules (0.1.0)
+
+- Users can customize the label of the file picker button.
+- Users can pass files in form of the data frame or reactive expression with
+  data.frame to the file picker module.